[refactor] Move literal construction to expr module (#4448)
* [refactor] Move literal construction to expr module

* Add errer message check

* Auto Format
