+++ tests/python/test_literal.py
@@ -23,7 +23,9 @@ def test_literal_multi_args_error():
     def multi_args_error():
         a = ti.i64(1, 2)
 
-    with pytest.raises(ti.TaichiSyntaxError):
+    with pytest.raises(
+            ti.TaichiSyntaxError,
+            match="Type annotation can only be given to a single literal."):
         multi_args_error()
 
 
@@ -33,7 +35,9 @@ def test_literal_keywords_error():
     def keywords_error():
         a = ti.f64(1, x=2)
 
-    with pytest.raises(ti.TaichiSyntaxError):
+    with pytest.raises(
+            ti.TaichiSyntaxError,
+            match="Type annotation can only be given to a single literal."):
         keywords_error()
 
 
@@ -44,5 +48,7 @@ def test_literal_expr_error():
         a = 1
         b = ti.f16(a)
 
-    with pytest.raises(ti.TaichiSyntaxError):
+    with pytest.raises(
+            ti.TaichiSyntaxError,
+            match="Type annotation can only be given to a single literal."):
         expr_error()
